refactor: ♻️ extract useEscapeKey hook to remove dialog duplication (#169)
